Move NoAuth implementation to separate file for consistency
