+++ crates/modelardb_auth/src/authenticator/no_auth.rs
@@ -13,25 +13,24 @@
  * limitations under the License.
  */
 
-//! Defines the [`Authenticator`] trait for validating credentials and authorizing access to
-//! ModelarDB.
-
-pub mod no_auth;
+//! An [`Authenticator`] that grants all permissions to all callers without any validation. Used in
+//! the open-source version of ModelarDB where authentication is not required.
 
 use tonic::Status;
 use tonic::metadata::MetadataMap;
 
 use crate::Permission;
+use crate::authenticator::Authenticator;
 
-/// Validates the credentials in `metadata` and checks that the caller has the `required_permission`.
-/// Implementations must return `Ok(())` if authorized, `Err(Status::unauthenticated(...))` if
-/// credentials are missing or invalid, or `Err(Status::permission_denied(...))` if credentials are
-/// valid but the caller lacks the required permission. Error messages must not reveal the specific
-/// reason for rejection to prevent information leakage.
-pub trait Authenticator: Send + Sync {
+/// An [`Authenticator`] that grants all permissions to all callers without any validation.
+pub struct NoAuth;
+
+impl Authenticator for NoAuth {
     fn authorize(
         &self,
-        metadata: &MetadataMap,
-        required_permission: Permission,
-    ) -> Result<(), Status>;
+        _metadata: &MetadataMap,
+        _required_permission: Permission,
+    ) -> Result<(), Status> {
+        Ok(())
+    }
 }
